pyproject.toml
[project]
name = "adcal2024"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "setuptools>=75.6.0",
    "solidpython>=1.1.3",
]
